+++ pkg/lib/saml/samlprotocol/error.go
@@ -1,4 +1,4 @@
-package protocol
+package samlprotocol
 
 import (
 	"fmt"
